# Simplify trimming non significant digits in JsonUtf8Writer